Avoid MetaMask dependency: Use eth-provider
This adds a new frontend dependency and connects to Infura to avoid
requiring MetaMask or any other local provider

This enabled lots of cleanup!

* Since we no longer need to use `window.ethereum` or `window.web3`, we no
  longer need to tell TypeScript about these attributes on `window`
* No more web3-enabled/disabled logic – both index.html and main.ts
  could be greatly simplified

One minor trade-off is that eth-provider does not have TypeScript
typings. I've added some minimal typings for it to `src/@types`, which
currently codes for the presets specified at:
https://github.com/floating/eth-provider/blob/e9ac97bf540b02bb90ebcbac8622a0a8a93fdc53/README.md#presets