Remove need for .map(...).join('')
When building UI out of arrays, I expect it to work a bit like React

Unfortunately, that's not how JavaScript works. When using a template
literal, when JavaScript encounters an array, it calls `toString()` on
it. The default `Array.prototype.toString()` joins the elements with
a comma, `this.join(',')`.

When a bunch of commas get inserted in between the elements we're trying
to add to the page, things look ugly. So we need to use `join('')` every
time we want to render an array.

Unless! We override the default `Array.prototype.toString`, and make it
join with empty string by default.

An alternative approach would be to use a tagged template, such as:

    html`<h1>stuff here</h1> ${myArray.map(...)}`

While I might reach for a tagged template in a larger app where globally
changing Array's default join behavior might cause frustration, this
app is small enough that I think think using the global override seems
more useful.