Walk Blocks: add placeholder block-walking logic
We want to find all _failed_ DAO creations, in addition to successful
DAO creations.

As far as our research has determined, the only way to do this is to
literally walk every single block from the creation block of our
contracts, checking transactions for ones against our contract.

😟 😨 😱

400,000 blocks to scan through. NBD.

This does not work, probably because of Infura rate limitations.

Where do we go from here?

It's probably time to try out The Graph, to see if we can aggregate
these stats with their backend, so this frontend can just grab the info
it needs over GraphQL.