[WIP] index data on The Graph using subgraph
Rather than parse 400,000 blocks in the user's browser and break
Infura's rate limiting, this was an attempt to move this
championship-level block parsing to a subgraph on thegraph.com

Unfortunately, The Graph's Ethereum API does not currently provide a way
to find failed transactions. I've requested such support here:
https://github.com/graphprotocol/support/issues/51

If we're going to only show successful DAO creations and info about each
of these, we may still want to go with The Graph, rather than process
all of this data on the user's machine. If we go that direction, we can
use `eventHandlers`, rather than `blockHandlers` or `callHandlers`.

* `callHandlers` seems cool, but does not work on Rinkeby, and we want
  Rinkeby stats
* `blockHandlers` seems to pretty limited, currently
* `eventHandlers` seems like the most common way to use The Graph, from
  looking through some example subgraphs